Split algorithm specific and generic parts for computing and saving/loading dimensionality reduction features from each other
